Rename index files to be more descriptive
